requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dotenv==1.1.0
requests==2.32.3
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
PyYAML
bcrypt==4.1.2
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
gunicorn==23.0.0
